+++ package.json
@@ -34,6 +34,8 @@
     "webpack-stream": "^5.2.1"
   },
   "dependencies": {
-    "jquery": "^3.4.1"
+    "jquery": "^3.4.1",
+    "moment": "^2.24.0",
+    "moment-range": "^4.0.2"
   }
 }
