+++ src/assets/data/holidays.yml
@@ -1,8 +1,8 @@
 ---
 -
     name: Новый год
-    start-date: 01-01
-    end-date: 01-01
+    start_date: 01-01
+    end_date: 12-01
     image: "/face/ny2018.gif"
     text: |-
         <h4>Дорогие читатели!</h4>
