+++ src/assets/js/utils.js
@@ -1,0 +1,3 @@
+export function getCurrentYear() {
+  return new Date().getFullYear(); // use _.once
+}
