+++ webpack.config.js
@@ -1,4 +1,5 @@
 const path = require("path");
+const webpack = require('webpack');
 
 module.exports = {
     entry: {
@@ -42,4 +43,8 @@ module.exports = {
             }
         ]
     },
+    
+    plugins: [
+      new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/),
+    ],
 };
